Fix ZTABLE key bugs
In using this for a project, or in testing it on the (npm tonic tester)[https://tonicdev.com/npm/ztable] (eg, with an input zscore of 1), I ran into an issue where several input z-scores would throw `'Uncaught TypeError: Cannot read property 'X' of undefined'` where X was an integer.

The issue, resolved in this PR, was that using `.toString()` on an integer would drop any decimals, and the `ZTABLE` keys all require one decimal place. Using `.toFixed(1)` instead resolves this issue.

On a related note, a `yZscore` of `0` would become the string `'0'` (`'0.0'` with this first edit), but never the `'-0.0'` key in the table. Therefore I edited this key to read `'0.0'` and provide a valid lookup.